feat: modify BaseOptions class to make it compatible with jupyer notebooks
For the parser to be used in jupyter notebooks, the command line arguments need to be manually put in a list and the passed to the parser.
This change allows the option to pass either a list or none to the parser.
If used within a script, None is passed to the parser and then the parser reads the command line args from the terminal.
If a list is passed, the parser reads the arguments from the list.

Before the change, None was the passed to the parser by default and thus it could not be used with a notebook.